+++ spec/features/sign_up_process_spec.rb
@@ -19,11 +19,4 @@ feature 'Visitor signs up' do
     expect(page).to have_content('Sign up')
   end
 
-  def sign_up_with(email, password)
-    visit new_user_registration_path
-    fill_in 'Email', with: email
-    fill_in 'Password', with: password
-    fill_in 'Confirm your password', with: password
-    click_button 'Sign up'
-  end
 end
