+++ spec/support/features.rb
@@ -1,0 +1,3 @@
+RSpec.configure do |config|
+  config.include Features::SessionHelpers, type: :feature
+end
